Convert to LLVM IR

+++ include/athena/backend/llvm/runtime/LaunchCommand.h
@@ -23,18 +23,18 @@ struct ArgDesc {
     TENSOR = 0,
     DATA = 1
   };
-  size_t size;
+  uint64_t size;
   void *arg;
   ArgType type;
 };
 
 struct LaunchCommand {
   const char* kernelName;
-  uint32_t argsCount;
+  uint64_t argsCount;
   ArgDesc *args;
-  size_t workDim;
-  size_t *globalSize;
-  size_t *localSize;
+  uint64_t workDim;
+  uint64_t *globalSize;
+  uint64_t *localSize;
 };
 
 #endif // ATHENA_LAUNCHCOMMAND_H
